third_party/blink/web_tests/external/wpt/css/css-grid/parsing/grid-auto-flow-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['grid-auto-flow'] = "row" should set the property value
PASS e.style['grid-auto-flow'] = "column" should set the property value
FAIL e.style['grid-auto-flow'] = "row dense" should set the property value assert_equals: serialization should be canonical expected "dense" but got "row dense"
FAIL e.style['grid-auto-flow'] = "dense row" should set the property value assert_equals: serialization should be canonical expected "dense" but got "row dense"
PASS e.style['grid-auto-flow'] = "dense" should set the property value
PASS e.style['grid-auto-flow'] = "column dense" should set the property value
PASS e.style['grid-auto-flow'] = "dense column" should set the property value
Harness: the test ran to completion.
